refactor: merge duplicated Protocol/StudyProtocol components
DeviceDropdown, ProtocolCards, and ProtocolInfo styles were duplicated
between pages/Protocol/ and pages/StudyProtocol/. DeviceDropdown and
ProtocolCards are now shared components in src/components/. The Protocol
page keeps a thin wrapper for data fetching and loading skeletons.
ProtocolInfo shared styles are extracted while keeping the two component
implementations separate since they serve different UX flows.
